Add VERSION pinning to the installation script
VERSION pinning is something that people have been asking for, for a
while so this attempts to add that in.

Should work in most scenarios aside from incomplete VERSIONS.

IE: If you have the `test` channel enabled and try to install `17.12`
for example, the script will instead try to install rc2 since that is
the latest VERSION in the results.

Should also work with nightly builds like pinning to a date, or a
specific git SHA

Signed-off-by: Eli Uriegas <eli.uriegas@docker.com>