Switch over to supported file
This ultimately gets rid of the SUPPORT_MAP variable. It was tough to
manage and wasn't entirely representative of what distributions were
currently supported on which channels.

This commit makes it so that we call out to a text file that tells us
which architectures and distributions are currently supported for each
`VERSION` including the `latest` version.

Signed-off-by: Eli Uriegas <eli.uriegas@docker.com>