Simplify CI, it was too complicated
Simplifies CI to run through a simple `make test` target that has an
underlying `TEST_IMAGE` variable which determines the distribution to
run on.

Also changes the Jenkinsfile to be declarative since the old one kind of
sucked.

Also adds a test for version pinning since that's something we rely on a
lot.

Signed-off-by: Eli Uriegas <eli.uriegas@docker.com>