Handle CentOS forks
Since CentOS 8 going into a rolling release mode with CentOS Stream, new forks have appeared for keeping the "stable release".

The first one to be available is Almalinux and the current script does not support it.

This PR addresses this gap, and as I don't yet the name/IDs of the future forks, I decided to go with a `case` statement, as it will be easier to manage in the long run.

By adding the `case` statement in the `get_distribution()` function, this avoids any impact on the creation of the distribution URL used in the script.

Signed-off-by: nunix corsair@hey.com