fix "REPO_FILE" checks always running, also with "dry run" enabled
These checks were originally added in 3217c543b4c0d54944ec4fb1133510aa8ea06eb7,
which added a `REPO_FILE` variable. While it's with good intent to do a check
if the URL is valid, there are some issues with these checks;

The checks depend on `curl`, but are executed _before_ `pre_reqs` are installed,
so will produce an error if `curl` is not found;

    ./install.sh
    # Executing docker install script, commit:
    ./install.sh: line 469: curl: command not found
    Error: Unable to curl repository file https://download.docker.com/linux/fedora/docker-ce.repo, is it valid?

The checks are currently always performed, also if `DRY_RUN=1` is enabled;

    DRY_RUN=1 REPO_FILE=no-such.repo ./install.sh
    # Executing docker install script, commit:
    Error: Unable to curl repository file https://download.docker.com/linux/fedora/no-such.repo, is it valid?

This patch removes the checks to address the above, and to simplify the script.
Running the script will still fail if an invalid URL is used for the `.repo` file,
which fails the script after the `pre_reqs` are installed;

With this patch:

When running the script with `DRY_RUN=1`:

    DRY_RUN=1 ./install.sh
    # Executing docker install script, commit:
    dnf install -y -q dnf-plugins-core
    dnf config-manager --add-repo https://download.docker.com/linux/fedora/docker-ce.repo
    dnf makecache
    dnf install -y -q docker-ce docker-ce-cli containerd.io docker-scan-plugin docker-compose-plugin docker-ce-rootless-extras

When running the script with `DRY_RUN=1` and an invalid repo, the script
completes successfully, but will show the commands it will run, including
the invalid URL:

    DRY_RUN=1 REPO_FILE=no-such.repo ./install.sh
    # Executing docker install script, commit:
    dnf install -y -q dnf-plugins-core
    dnf config-manager --add-repo https://download.docker.com/linux/fedora/no-such.repo
    dnf makecache
    dnf install -y -q docker-ce docker-ce-cli containerd.io docker-scan-plugin docker-compose-plugin docker-ce-rootless-extras

When running the script with an invalid repo URL, it will fail early (after
installing `pre_reqs`:

    REPO_FILE=no-such.repo ./install.sh

    # Executing docker install script, commit:
    + sh -c 'dnf install -y -q dnf-plugins-core'
    ...

    + sh -c 'dnf config-manager --add-repo https://download.docker.com/linux/fedora/no-such.repo'
      Adding repo from: https://download.docker.com/linux/fedora/no-such.repo
      Status code: 404 for https://download.docker.com/linux/fedora/no-such.repo (IP: 65.9.86.56)
      Error: Configuration of repo failed

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>