deb: remove --no-install-recommends
The installation for deb packages used a --no-install-recommends to skip "recommended"
packages. This flag was added in ab5553e3027fdf53c5b68ef543f7437dc15c33df to work around
an issue on Ubuntu 17.10 (which had aufs-tools as recommended dependency), causing the
kernel to be updated.

While "recommended" packages _can_ be omitted during installation, they are
a "strong" recommendation, and should really be installed by default:

> Recommends
>
> This declares a strong, but not absolute, dependency.
> The Recommends field should list packages that would be found together
> with this one in all but unusual installations.

Including the `--no-install-recommends` option could result in `apparmor_parser`
binary (part of the `apparmor` package) to be missing, which would cause
the daemon failing to start on systems with AppArmor support enabled:

    Error response from daemon: AppArmor enabled on system but the docker-default profile could not be loaded: running `apparmor_parser apparmor_parser --version` failed with output:
    error: exec: "apparmor_parser": executable file not found in $PATH
    Error: failed to start containers: somecontainer

This patch removes the `--no-install-recommends` so that the recommended
packages are installed.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>