SLES: remove unused rootless_pkg_version
This variable was introduced in 63fb2060c9436bb3ad4434bdbf55f92824428e8e, but
was later refactored in ab0389264936174e064d0fcf3ab39d2ee8d9b976 aligned the
installation steps for SLES with the other distros (using the version of the
"docker-ce" package itself, as those versions always match).

This patch removes the unused rootless_pkg_version variable, and the code to
resolve that version.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>