remove redundant install for scan-cli-plugin
This explict install was added because docker 20.10 did not have an explicit
dependency on the plugin. The current release of the Docker CLI (23.0) already
has a "Recommends:" dependency on the plugin, and 0095742fb5d923f876e720731073c541f47f0f79
removed the "--no-install-recommends", so fresh installs of Docker 23.0 will
already install this plugin through that dependency.

Removing the explicit install, because the scan-cli-plugin has been deprecated,
and will no longer receive updates, so we don't want it to be installed for the
upcoming 24.0 release when using this script. see;
https://github.com/docker/scan-cli-plugin/blob/b69ac460a6c23243b858326f0e26f6bd202ca781/internal/deprecation.go#L13-L15

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>