update script to use .asc, remove gpg as pre-requisite
Current distro versions now allow the GPG key to be stored as an `.asc` file,
which simplifies installation. Changing the file extension for the GPG key
to .asc lets apt-get know the GPG key is in ASCII armor format, in which case
there is no need to "dearmor", and we no longer need gpg as a pre-requisite.
Instructions for manual installation on Debian based distros were updated in
our documentation through [docker/docs@33befd6].

It's common for users to install packages in a combination of the convenience
script and manual (e.g. to update the packages after install, or vice-versa).
Given that both the script and the manual steps configure the repository, we
want those steps to align, otherwise (potentially?) warnings / errors may occur;

    $ sudo apt-get update
    E: Conflicting values set for option Signed-By regarding source https://download.docker.com/linux/debian/ bullseye: /usr/share/keyrings/docker-ce-archive-keyring.gpg != /usr/share/keyrings/docker-archive-keyring.gpg
    E: The list of sources could not be read.

In either case; let's make sure that our script aligns with the manual steps
that we document.

[docker/docs@33befd6]: https://github.com/docker/docs/commit/33befd692249d11ce7bc88052d82709a07a8107f

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>