Makefile: Ignore SC2329
In previous versions of shellcheck, SC2317 was emitted for every line of
a unused function and we're ignoring the SC2317 completely:
https://github.com/docker/docker-install/blob/6e61729e66af62fd54c0b659751a431be8f90f68/Makefile#L6

We were ignoring this error due to `version_compare` being reported as a
false positive.

After
https://github.com/koalaman/shellcheck/commit/4f81dbe839091a06a5cfaea695cf1c451ff07565
this was fixed and now it produces one `SC2329` for the whole function
instead of multiple SC2137.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>