remove check for RHEL only being available on s390x
This check was added in 63fb2060c9436bb3ad4434bdbf55f92824428e8e, at which
time we did not provide packages fro RHEL on architectures other than s390x.

We now provide packages for RHEL 8 and 9 for amd64 (x86), arm64, and s390x,
so we can remove this check.

Note that we do not (yet?) provide packages for ppc64le, but this is more
of a niche use-case, so not adding a special condition for that, as we
don't provide that check for other distros.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>