Avoid installing missing packages for version 23
On ubuntu bionic and focal (and debian buster too), the docker-model-plugin
package is missing. The suggestion was to just install the package named
docker-buildx-plugin in the version 23 check, while moving the missing
package docker-model-plugin to a new version 28.2 check. The number was
chosen because that package becomes available starting with version 28.2
on ubuntu jammy (while not available in version 28.1 on ubuntu focal).

See: https://github.com/docker/docker-install/issues/509

Signed-off-by: Rejean Groleau <rejeangroleau@gmail.com>